Rename kinetics reader modules
